requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
databases[postgresql]==0.8.0
